para topo 24-09-2014

+++ SOP_IAA_DAL/ClassAnnotations/ProgProyAnnotations.cs
@@ -28,7 +28,7 @@ namespace SOP_IAA_DAL
 
         [Required]
         [DisplayName("Monto Estimado")]
-        public int monto { get; set; }
+        public decimal monto { get; set; }
 
   
     }
